pyproject.toml
[tool.poetry]
name = "datawarehouse"
version = "0.1.0"
description = ""
authors = ["Deborah Foroni <deborah.foroni@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
briefer = "^0.0.14"
dbt-postgres = "^1.9.0"
streamlit = "^1.41.0"
pydantic = "^2.10.3"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
